api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.116.1",
    "gpx2mesh>=0.1.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
